✨ feat(login): add auto_login parameter
Adds an `auto_login` parameter to `TeamTalkBot.add_server`, `enum.TeamTalkServerInfo` and `TeamTalkInstance` to allow manual control over the login process.

When `auto_login` is set to `False`, the bot will connect to the server but will not automatically log in. This is useful for scenarios that require a manual login step, such as logging in with a BearWare.dk account.

The default value is `True` to maintain the existing behavior of automatically logging in after a successful connection.

Signed-off-by: Patrick ZAJDA <patrick@zajda.fr>